+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- Declarative runbook controls for step and stage policies.
+
 ## 0.1.0 - 2026-05-09
 
 Initial release.